cloud_function/functions/requirements.txt
firebase_functions~=0.1.0
numpy
faiss-cpu
pandas
